Cleaning up the ml_utils and utils modules
